Fixed stack issue so the system behaves OK by now

+++ FreeRTOS_Shell/sources/sources/libs/ao_config/inc/ExtiConfig.hpp
@@ -11,7 +11,7 @@ struct ExtiConfig {
     uint32_t            extiLine;   // EXTI12, EXTI0, etc.
     exti_trigger_type   trigger;    // EXTI_TRIGGER_FALLING / BOTH
     uint8_t             nvicIrq;    // NVIC_EXTI15_10_IRQ, etc.
-    uint8_t             nvicPrio;
+    uint32_t            nvicPrio;
     uint8_t             lineNumber; // numeric 0-15, used by registry
 };
